Navbar toggle on "touchscreen" media query, split up components into files
