+++ cerg-app/package.json
@@ -33,8 +33,9 @@
     "@types/react": "18.0.9",
     "@types/react-dom": "18.0.3",
     "@types/react-slick": "^0.23.8",
-    "eslint": "8.15.0",
+    "eslint": "^8.15.0",
     "eslint-config-next": "12.1.6",
+    "eslint-plugin-react": "^7.30.0",
     "typescript": "4.6.4"
   }
 }
