+++ cerg-app/pages/Ann-Perry%20Witmer.tsx
@@ -1,5 +1,5 @@
 import Footer from "../src/components/footer";
-import NavBar from "../src/components/navbar";
+import NavBar from "../src/components/NavBar/NavBar";
 import { PEOPLE } from "../src/types/team-members";
 import BioPage from "./bio-page";
