+++ cerg-app/pages/contact.tsx
@@ -1,6 +1,6 @@
 
 import Footer from "../src/components/footer";
-import NavBar from "../src/components/navbar";
+import NavBar from "../src/components/NavBar/NavBar";
 
 export default function TemplatePage() {
   return (
