+++ cerg-app/pages/error.tsx
@@ -1,6 +1,6 @@
 import { Container, Heading } from "@chakra-ui/react";
 import Footer from "../src/components/footer";
-import NavBar from "../src/components/navbar";
+import NavBar from "../src/components/NavBar/NavBar";
 
 export default function ErrorPage() {
   return (
