+++ cerg-app/pages/focus-areas.tsx
@@ -1,6 +1,6 @@
 import { Flex, Spacer } from "@chakra-ui/react";
 import Footer from "../src/components/footer";
-import NavBar from "../src/components/navbar";
+import NavBar from "../src/components/NavBar/NavBar";
 
 export default function FocusAreasPage() {
   return (
