+++ cerg-app/pages/news.tsx
@@ -1,5 +1,5 @@
 import Footer from "../src/components/footer";
-import NavBar from "../src/components/navbar";
+import NavBar from "../src/components/NavBar/NavBar";
 import WithBackgroundImage from "../src/demo-components/with-background-image-and-gradient";
 
 export default function NewsPage() {
