+++ cerg-app/pages/research.tsx
@@ -1,6 +1,6 @@
 import Footer from "../src/components/footer";
 import Hero from "../src/components/hero";
-import NavBar from "../src/components/navbar";
+import NavBar from "../src/components/NavBar/NavBar";
 
 export default function ResearchPage() {
   return (
