+++ cerg-app/pages/resources.tsx
@@ -1,5 +1,5 @@
 import Footer from "../src/components/footer";
-import NavBar from "../src/components/navbar";
+import NavBar from "../src/components/NavBar/NavBar";
 
 export default function ResourcesPage() {
   return (
