+++ cerg-app/pages/team.tsx
@@ -1,6 +1,6 @@
 import { Box, Center, Flex, Heading, VStack } from "@chakra-ui/react";
 import Footer from "../src/components/footer";
-import NavBar from "../src/components/navbar";
+import NavBar from "../src/components/NavBar/NavBar";
 import BioCard from "../src/components/bio-card";
 import { PEOPLE } from "../src/types/team-members";
 import { useState } from "react";
