+++ cerg-app/pages/template.tsx
@@ -1,5 +1,5 @@
 import Footer from "../src/components/footer";
-import NavBar from "../src/components/navbar";
+import NavBar from "../src/components/NavBar/NavBar";
 
 export default function TemplatePage() {
   return (
